Finish d25 and express stage

+++ d25_express_zadanie_podsumowujace/router/client.js
@@ -1,38 +1,55 @@
 const express = require('express');
 const clientRouter = express.Router();
 const db = require('../utils/db');
-const {ClientRecord} = require('../records/client-record');
+const {NotFoundError} = require("../utils/errors");
+
 
 clientRouter
 
     .get('/', (req, res) => {
-        // res.json(db.getAll());
         res.render('client/list-all', {
             clients: db.getAll(),
         });
     })
 
     .get('/:id', (req, res) => {
-        const myClient = new ClientRecord(db.getOne(req.params.id));
-        console.log(myClient);
+        const client = db.getOne(req.params.id);
 
-        const {id} = req.params;
+        if (!client) {
+            throw new NotFoundError();
+        }
+
         res.render('client/one', {
-            client: db.getOne(id)
+            client
         });
     })
 
-    .post('/', async (req, res) => {
-        const {name, mail, notes, nextContactAt} = req.body;
-        const id = await db.create({name, mail, notes, nextContactAt});
-        res.render('client/added', {
-            name: req.body.name,
-            id,
-        });
+    .post('/', async (req, res, next) => {
+        try{
+            // asdasdasd();
+            const {name, mail, notes, nextContactAt} = req.body;
+            const id = await db.create({name, mail, notes, nextContactAt});
+            res
+                .status(201)
+                .render('client/added', {
+                name: req.body.name,
+                id,
+            });
+        } catch (error) {
+            next(error);
+        }
+
     })
 
     .put('/:id', async (req, res) => {
         const {id} = req.params;
+
+        const client = db.getOne(id);
+
+        if (!client) {
+            throw new NotFoundError();
+        }
+
         await db.update(id, req.body);
         res.render('client/modified', {
             name: req.body.name,
@@ -42,6 +59,13 @@ clientRouter
 
     .delete('/:id', async (req, res) => {
         const {id} = req.params;
+
+        const client = db.getOne(id);
+
+        if (!client) {
+            throw new NotFoundError();
+        }
+
         await db.delete(id);
         res.render('client/deleted');
     })
@@ -51,9 +75,14 @@ clientRouter
     })
 
     .get('/form/edit/:id', (req, res) => {
-        const {id} = req.params;
+        const client = db.getOne(req.params.id);
+
+        if (!client) {
+            throw new NotFoundError();
+        }
+
         res.render('client/forms/edit', {
-            client: db.getOne(id)
+            client
         });
     });
